refactor : Change compute directory structure
