+++ Sources/Motutapu/compute/Sparse.cpp
@@ -4,9 +4,9 @@
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
-#include <Motutapu/compute/Sparse.hpp>
-#include <Motutapu/compute/SparseMatrix.hpp>
-#include <Motutapu/compute/cuda/Memory.hpp>
+#include <Motutapu/compute/cudaUtil/Memory.hpp>
+#include <Motutapu/compute/sparse/Sparse.hpp>
+#include <Motutapu/compute/sparse/SparseMatrix.hpp>
 #include <Motutapu/util/MemoryManager.hpp>
 
 namespace Motutapu::Compute
